refactor(gui): extract update front door module
